Bake failure

+++ src/code/control/TestSuite.java
@@ -53,7 +53,7 @@ public class TestSuite {
         callback.addIngredient(ingredient);
         callback.incrementIngredient(ingredient.getType());
 
-        int nIngredients = callback.getNumIngredients(ingredient.getType());
+        double nIngredients = callback.getNumIngredients(ingredient.getType());
         assertEquals(nIngredients, 2);
     }
 
@@ -71,7 +71,7 @@ public class TestSuite {
         callback.incrementIngredient(ingredient.getType());
         callback.decrementIngredient(ingredient.getType());
 
-        int nIngredients = callback.getNumIngredients(ingredient.getType());
+        double nIngredients = callback.getNumIngredients(ingredient.getType());
         assertEquals(nIngredients, 1);
     }
