+++ src/code/entities/Ingredient.java
@@ -12,7 +12,7 @@ public class Ingredient implements Serializable {
     private transient static double serialVersionUID = 51D;
     private String type, category;
     private Supplier supplier;
-    private int stock;
+    private double stock;
     private Units unit;
     private String stockAndUnit;
 
@@ -23,7 +23,7 @@ public class Ingredient implements Serializable {
         stock = 1;
     }
 
-    public Ingredient(String type, String category, int stock, Supplier supplier, Units unit) {
+    public Ingredient(String type, String category, double stock, Supplier supplier, Units unit) {
         this.type = type;
         this.category = category;
         this.supplier = supplier;
@@ -56,11 +56,11 @@ public class Ingredient implements Serializable {
         this.supplier = supplier;
     }
 
-    public int getStock() {
+    public double getStock() {
         return stock;
     }
 
-    public void setStock(int stock) {
+    public void setStock(double stock) {
         this.stock = stock;
     }
